+++ src/entityLogoCache.hpp
@@ -38,10 +38,11 @@ public:
 	
 	static EntityLogoCache& getInstance() noexcept;
 	
+	virtual QImage getImage(la::avdecc::UniqueIdentifier const entityID, Type const type, bool const downloadIfNotInCache = false) noexcept = 0;
+	virtual bool isImageInCache(la::avdecc::UniqueIdentifier const entityID, Type const type) const noexcept = 0;
+
 	virtual void clear() noexcept = 0;
-	
-	virtual QImage getImage(la::avdecc::UniqueIdentifier const entityID, Type const type, bool const forceDownload = false) noexcept = 0;
-	
+
 	Q_SIGNAL void imageChanged(la::avdecc::UniqueIdentifier const entityID, EntityLogoCache::Type const type);
 	
 protected: