+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to Hive will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
 ## [1.0.3] - 2018-07-17
 ### Added
 - Support for Memory Object descriptors (Contributed by Florian Harmuth)