+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [Display entity's logo image in the entity list](https://github.com/christophe-calmejane/Hive/issues/10)
+- Support for manufacturer and entity images download in Memory Object descriptors
 
 ## [1.0.3] - 2018-07-17
 ### Added